boost: moveonly: split CPubKey and friends to new files
